refactor history tables
